remove locks on sender-side of mpsc channel
Remove the locks on the shared app2client Sender. Senders implement the Send Trait. This allows them to be shared by copying and moving them to other threads (for example, see the "Shared Usage" example in
https://doc.rust-lang.org/std/sync/mpsc).

This simplifies the function calls and the unpacking logic. Furthermore, this reduces overall locking, as the Sender internally takes care of synchronization.